refactor/data_packaging: clean up test imports and make test pass
